+++ src/main/java/GameWindow.java
@@ -1,16 +1,7 @@
-import com.thoughtworks.xstream.XStream;
-
 import javax.swing.*;
-import javax.swing.filechooser.FileNameExtensionFilter;
-import javax.swing.filechooser.FileSystemView;
 import java.awt.*;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.Arrays;
-
 //import java.time.LocalDateTime;
 
 
@@ -86,10 +77,11 @@ public class GameWindow {
                     connectionHandler.stopReceiving();
             }
         });
-        saveGameButton.addActionListener(e -> {
+       /* saveGameButton.addActionListener(e -> {
             if (checkIfBoardWasAltered())
                 saveGame(); // save current game
-        });
+        });*/
+       saveGameButton.addActionListener(new ActionSaveAs(this));
         quitGameButton.addActionListener(e -> {
             if (!checkIfBoardWasAltered() || ignoredWarning("Do you want to exit game without saving?")) {
                 System.exit(0);
@@ -105,56 +97,6 @@ public class GameWindow {
         window.setJMenuBar(menuBar);
     }
 
-    private void saveGame() {
-        JFileChooser fileChooser = createFileChooser();
-        fileChooser.setSelectedFile(new File("chessGame_" + java.time.LocalDate.now()));
-/*
-        System.out.println("Project Directory : "+ System.getProperty("user.dir"));
-        String gameDirectory = System.getProperty("user.dir");
-        jfc.setSelectedFile(new File(gameDirectory + "/" + "chessGame_" + java.time.LocalDate.now())); // sets game folder as default to save file in
-*/
-        int returnValue = fileChooser.showSaveDialog(null);
-        if (returnValue == JFileChooser.APPROVE_OPTION) {
-            File selectedFile = fileChooser.getSelectedFile();
-            if (!selectedFile.exists() || ignoredWarning("Do you want to overwrite data in file?")) { // prevents overwriting existing file without player consent
-                String fileName = getFileName(fileChooser);
-                //
-                // save game state to xml file /////////////////////////////
-
-                saveXmlGame(fileName);
-
-                //if(successful save)
-                board.setBoardAltered(false);
-                //
-            }
-        }
-    }
-
-    public JFileChooser createFileChooser() {
-        // fileChooser parameters
-        JFileChooser fileChooser = new JFileChooser(FileSystemView.getFileSystemView().getHomeDirectory());
-        fileChooser.setDialogTitle("Choose a folder and file name to save game: ");
-        //jfc.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
-        fileChooser.setAcceptAllFileFilterUsed(false); // restricts file formats to those declared below
-        FileNameExtensionFilter filterXML = new FileNameExtensionFilter("XML files", "xml");
-        fileChooser.addChoosableFileFilter(filterXML);
-        return fileChooser;
-    }
-
-    public String getFileName(JFileChooser fileChooser) {
-        File selectedFile = fileChooser.getSelectedFile();
-        String selectedFileName = selectedFile.getAbsolutePath();
-
-        System.out.println("Player has chosen to save game at: " + selectedFileName);
-        FileNameExtensionFilter fileFilter = (FileNameExtensionFilter) fileChooser.getFileFilter();
-        System.out.println("Accepted extensions: " + Arrays.toString(fileFilter.getExtensions()));
-        if (!fileFilter.accept(selectedFile))
-            selectedFileName = selectedFileName + "." + fileFilter.getExtensions()[0];
-        System.out.println("Game will be saved at: " + selectedFileName);
-
-        return selectedFileName;
-    }
-
     public JFrame createWarningWindow() {
         final JFrame warningWindow = new JFrame("Warning");
         warningWindow.setSize(new Dimension(300, 100));
@@ -163,7 +105,7 @@ public class GameWindow {
         return warningWindow;
     }
 
-    private boolean ignoredWarning(String warningMessage) {
+    public boolean ignoredWarning(String warningMessage) {
         JFrame warningWindow = createWarningWindow();
         boolean yesResponse = JOptionPane.showConfirmDialog(warningWindow, warningMessage) == JOptionPane.YES_OPTION;
         warningWindow.setVisible(false);
@@ -187,30 +129,11 @@ public class GameWindow {
         this.opponentName = opponentName;
     }
 
-    public boolean checkIfBoardWasAltered() {
-        return board.checkIfBoardAltered();
-    }
-
     public Board getBoard() {
         return board;
     }
 
-    public void saveXmlGame(String filename) {
-        SavedGame game = new SavedGame(this);
-
-        XStream xstream = new XStream();
-        System.out.println(xstream.toXML(game));
-        String xmlGame = xstream.toXML(game);
-
-        // write xml string to the specified file
-        try {
-            FileWriter fileWriter = new FileWriter(filename);
-            fileWriter.write(xmlGame);
-            fileWriter.close();
-            System.out.println("Successfully wrote to the file.");
-        } catch (IOException e) {
-            System.out.println("Error occurred while saving game to xml file");
-            e.printStackTrace();
-        }
+    public boolean checkIfBoardWasAltered() {
+        return board.checkIfBoardAltered();
     }
 }
